move load_all_configs to paasta_tools.util.configs
