+++ paasta_tools/cleanup_kubernetes_cr.py
@@ -31,8 +31,8 @@ from paasta_tools.kubernetes_tools import KubeClient
 from paasta_tools.kubernetes_tools import list_custom_resources
 from paasta_tools.kubernetes_tools import load_custom_resource_definitions
 from paasta_tools.kubernetes_tools import paasta_prefixed
+from paasta_tools.util.configs import load_all_configs
 from paasta_tools.utils import DEFAULT_SOA_DIR
-from paasta_tools.utils import load_all_configs
 from paasta_tools.utils import load_system_paasta_config
 
 log = logging.getLogger(__name__)
