+++ paasta_tools/utils.py
@@ -3662,19 +3662,6 @@ def get_possible_launched_by_user_variable_from_env() -> str:
     return os.getenv("SUDO_USER") or getpass.getuser()
 
 
-def load_all_configs(
-    cluster: str, file_prefix: str, soa_dir: str
-) -> Mapping[str, Mapping[str, Any]]:
-    config_dicts = {}
-    for service in os.listdir(soa_dir):
-        config_dicts[
-            service
-        ] = service_configuration_lib.read_extra_service_information(
-            service, f"{file_prefix}-{cluster}", soa_dir=soa_dir
-        )
-    return config_dicts
-
-
 def _reorder_docker_volumes(volumes: List[DockerVolume]) -> List[DockerVolume]:
     deduped = {
         v["containerPath"].rstrip("/") + v["hostPath"].rstrip("/"): v for v in volumes
